requirements.txt
flask>=2.3
flask-cors>=3.0.10
torch>=2.2.0
torchvision>=0.17.0
efficientnet_pytorch>=0.7.1
Pillow>=9.5.0
